Add Detector Attach/Detach
- also modified some HTTP response codes

A one-to-one relationship between Locations and Detectors has been
defined. Both entity types are "self-sufficient", so there's no cascade
delete, e.g. if you delete a Location, the Detector attached to it does
not get deleted, only its "Location" reference is reset.

Detectors can't be deleted when they're busy (i.e. in the "Running"
state).

Locations can't be deleted, if the Detector associated with them is
busy.