Use DbContextFactory in DetectorController
Reasoning:

By default, DbContext lifetime is scoped, which means a new instance of
DbContext is created for each "connection", like an HTTP query
(short-lived request) or a WebSocket connection between the backend and
a detector (typically a long-lived request). DbContexts are not
thread-safe, and different DbContext instances are not aware of the
changes made to the tracked entities by other instances. This can cause
entities tracked by contexts created in long-lived requests to become
outdated.

DbContextFactory is used to create a new DbContext instance before
processing a Command in DetectorController. We create the context
**right after** the command is dequeued to ensure that it is fresh (the
service might block for a long time on EnqueueEvent). Separate contexts
are created in RegisterDetector and DisconnectDetector for the same
reason.