Add CRUD operations (except Update) to Jobs
... also extended some (not yet used) domain entities and made some
non-breaking changes to existing endpoints

Jobs shouldn't be deleted or updated while they're "active", i.e. have
child Tasks that are in the Active state, or the detector connected to
the location is in the Running state

The Update endpoint should only be concerned with updating the Name and
Type properties. The relationship between Jobs and Tasks is 1-N, so the
Jobs table doesn't actually contain references to the child Tasks.