Revert to using manual model validation
Reasoning:
1. https://docs.fluentvalidation.net/en/latest/aspnet.html
2. Automatic model validaton was causing all sorts of issues that I was
   unable to resolve in a couple hours. Unexplicable HTTP 500 return
   codes were the most common (string<->enum conversions, missing
   properties or invalid JSON).

For now, model validation is performed by calling the
ValidateToModelStateAsync extension method on the injected validator. If
the validation fails on any of the specified rules, a 400 Bad Request is
returned (actually an instance of ValidationProblem is returned, for
consistent formatting of auto model binding and custom errors).

For the Update endpoints, all fields in the Request should be required.
Unfortunately, when a value type property is missing from the request,
the model binder initializes that property to its default value. A way
around this is to mark the property as nullable, in which case it's
initialized to null. We must then assign a NotNull rule to the property
to mark it as required.