Monitoring commands, BE->processor communication + some refactoring (huge commit)
=== Monitoring commands: (StartMonitoring/StopMonitoring)
The commands aren't forwarded to the detectors and are handled by the
backend instead. For this reason, these operations are tied to
Locations, and are not DetectorCommand variants
(StartMonitoring: ~start monitoring Task T on Location L~)
Monitoring is included as a Detector state, which isn't exactly right,
because from the detector's point of view, there's no difference between
Streaming and Monitoring (they just spit out camera frames in both cases).

=== Backend -> Processor communication:
Processor: python process that runs in parallel with the backend.
Responsible for running AI algorithm on inputs from backend, and
returning the results

Implemented using Unix sockets, available on Linux and newer versions of
Windows 10+. Unix sockets only give us a byte stream between the two
processes, and we need to build a binary protocol on top of it.
This protocol is only halfway done (Processor->BE communication
is not yet defined).

The basic ideas:
 - the format of packets should be well defined and the ordering of
   their properties should be fixed
 - integers should be 4 bytes wide
 - for enums, integer representation is used (either 1 byte or 4 bytes,
   haven't really decided yet, but it doesn't matter too much)
 - different types of messages are distinguished by the first (few)
   byte(s) of the message
 - for data of indeterminate size (strings, lists, JPEG images), the
   data should be preceded by an integer containing the byte size of the
   data that directly follows it

This way, the client does not have to know the exact size of each
packet before reading it from the stream, and there is no ambiguity
concerning the meaning of the data being read.

There are 2 types of packets in the BE -> processor direction:
(naming isn't great any will probably change in the future)
1. Request: the camera frame that the processor has to process
 - detector id: for choosing which set of parameters to use
 - frame size: for how many bytes of image data to read
 - frame

2. Params: additional parameters necessary for the algorithm to run
 - detector id: serves as an ID
 - job type: for which algorithm to run
 - templates count: for how many templates to read
 - templates (id, x, y, w, h)

When Monitoring is started, Params are sent to the processor. Each
subsequent camera frame captured by the detector on the specific
location will be wrapped in a Request packet, and sent to the processor.

Refactoring:
 - renamed some classes to better reflect their role
 - abbreviations (Extension->Ext, Exception->Ex, Options->Opt)
 - used type aliases instead of primitives where it helps understanding
   (e.g. DetectorId instead of int)