Processor: read frame packets + some refactoring
Fixed a bug caused by my (dumb and insignificant) premature
optimization: in StreamHandler on the backend, the DetectorFrameBuffer
was originally a struct (value type) instead of a class (reference
type). When retrieving an instance of the frame buffer from the
dictionary, its value was copied (because it's a value type), and the
modified value was never written back. The backend ended up sending
basically empty frames to the processor