Run BE<->Processor communication through 2 sockets
This is to avoid the socket deadlocking when the socket buffer gets
full. The issue was that the Processor was not able to keep up with the
speed at which the backend had sent the frame packets, and when the
socket's backing buffer got full, it blocked on sending results back to
the backend, so it was unable to get back to reading. I'm not familiar
enough with IPC socket internals to know exactly why this happens.

I experimented with using the `multiprocessing` module to speed up the
Processor, but I couldn't wrap my head around how sharing complex
objects (e.g. the frame queue) between processes works.

Instead the solution I chose was to use 2 sockets in parallel, one for
each direction (the `req` socket for BE->Processor, the `res` socket for
Processor->BE). On the Processor end, the main (receiver) thread and the
Runner (sender) thread(s) don't need to share the same socket anymore,
so deadlocking never occurs (+ the receiver can keep up with the rate at
which the BE sends packets for the same reason).