Processor: fix Queue impl bug
This made the receiver not to skip over the queue slot that was locked
by the Runner, which caused the receiver to stagger.

This also made the BE send corrupted frames (PacketSender.SendFrame is
not expected to block). Fixing the bug got rid of this, but it should
still be looked into because it's wrong