TaskInstance creation logic & share TaskId with Processor
TaskId is given to the Processor in ParamsPacket, and it's received back
in subsequent ResultPackets from the same Runner. This lets us skip a
slow and ugly database query in PacketReceiverService.ProcessResult
(getting active TaskInstance from the detector id). DetectorId is still
used internally by the Processor for grouping incoming FramePackets,
which do not contain the TaskId (they could if they performed the same
ugly detector id -> active task query on every incoming frame - although
caching could be used to alleviate that)