Add StopPacket
When StopMonitoring is requested for a location, or when a task
completes, the processor should stop sending ResultPackets to the BE.
When a processor receives a StopPacket, it pauses the specified Runner.

An alternative approach would be to have a "Processed" flag on each
Queue slot, which is set when the slot is read, and reset when the slot
is written. When monitoring is stopped, the task's Runner stops
receiving frames, and it would run out of unprocessed slots. There'd be
a need for some additional logic for restarting the Runner.