BE: Add flexible database seeding from JSON
All *.json files living in SeedData/ are processed at startup
by SeedLoader.LoadFromJson. The only way of doing this is in a generic
way is by relying heavily on C# reflection.

This is kind of an experiment and definitely over-engineered, but it's
also cool for a couple reasons:
 - ability to define seed data in JSON (comfy) vs. directly in the code
   (not so comfy)
 - ability to have a set of initial data that is version controlled,
   plus each developer can have their own, additional set of data that
   is not version controlled

The json files must have the following structure:

{
    "<entity plural name>": [
      {
	  "<entity property1 name>": <entity property1 value>,
	  "<entity property2 name>": <entity property2 value>,
	  "<entity property3 name>": <entity property3 value>,
	  <...>
      },
    ],
    <...>
}

Parsing the seed data works as you'd expect. For property types that
don't have a JSON equivalent (e.g. Detector.MacAddress) or data that
can't easily be expressed in JSON (e.g. Job.Snapshot), custom parsers
are used, which are instantiated in SeedLoader.SpecialParsers (e.g.
snapshots are parsed from files residing in the SeedData/Files/
directory).

For relationships, setting the foreign key property on one side of the
relation is generally enough for EF Core to wire stuff up. Navigation
properties are left blank. Unfortunately, the entity IDs have to be
specified by the user in order for relationships to be definable (there
might be a way around this).

As before, seeding is only performed in Development mode, and only if
the database does not exist when the backend is started up. In
Production mode, only the empty database is created.