Processor: allow dummy detector to read results from script
When invoking the Processor, you can add an argument containing a file
name of a "dummy script" file, from inside the dummy_scripts/ folder.
The script files specify the states of templates (in case of kit tasks),
or QA results (in case of QA tasks) along a timeline. When the runner is
started up, it will begin traversing his timeline from the beginning
with the given period, and report the current set of results at that
specific point.

The script files are in JSON format. Check the examples in
dummy_scripts/ if you're interested.