FE: Fix *Card tooltip behavior when clicked
Previously the tooltips briefly jumped to the upper left corner before
disappearing, when clicking one of the action buttons on a LocationCard
/ DetectorCard component. This is because the area containing the action
buttons for a given card is only visible when the card is hovered. As
soon as the button is clicked, focus is transfered to the popup window,
and the area disappears. The 'followCursor' attribute on the Tooltip
fixes this by gluing the tooltip to the cursor instead