BE: Split Templates and StateChanges in data model
 + removed OrderNum from ParamsPacketTemplate (unused)

When editing kit tasks on the interface, you need to select an area
for each template, even if the templates are supposed to represent
state changes for the same physical object. This works in practice, but
it could be logically cleaner. The solution is to split the Template
entity into two parts: one that references the physical object itself
(Template) + one that defines the expected change in state for the
object (StateChange). We may define multiple expected state changes for
objects in any given task, so the Template->StateChange relationship is
1-N.

The task creation workflow will change slightly. First, users must
select templates on the task snapshot that represent the objects they
wish to monitor. Then, the expected state changes must be defined. Each
expected state change must be connected to a template.

Untested!