FE: Basic InstancePanel & InstanceCard components
Will have a single list of cards instead of the wireframed layout
('Current' tab for events of currently running task instance + 'All' tab
for all task instances on location). Each cards represents a
TaskInstance. This is cleaner and probably more straightforward.

Cards are sorted by creation date (newest on top).

Each TaskInstance card can be expanded to display the contained Events
(no popup like in the wireframe). If a TI is running, its card is
highlighted, expanded, and on top (it's guaranteed to be the newest one
- new TIs can only be created after the previous one is finished). This
ensures that Events related to the currently running TI can be viewed
while on the dashboard.