Make generic company hierarchy handling work
... by removing the AddChildNode declaration from the interface. This
isn't great as the signature & implementation are basically the same
across all levels and could be extracted into a default impl in the
interface.

The private parameterless constructors are needed on the domain entities
by EF Core and the seed loader. As they're private, they don't violate
the domain boundary.