Implement Command binding & sending
The DetectorCommand is an abstract type with each concrete Command type
implementing it. This is done this way because it allows for nicely
modeling command parameters, e.g. taskId on StartDetection. Other
options would be using a single DetectorCommand concrete type with
nullable fields for each parameter (gets ugly quickly), or with a
generic JSON property.

To bind the request data in the body of a Detectors.Command request to a
DetectorCommand instance, a custom request binder is used
(CommandReqBinder).

All command types and snapshot and streaming requests are sent to
detectors through a DetectorHttpConnection instance. 127.0.0.1 is used
as a temporary IP address until IPs are saved by the Identify endpoint.